@@ -1,14 +1,14 @@
 ﻿using AutoMapper;
 using InventoryItems.Domain.Exceptions;
 using InventoryItems.Domain.Interfaces.Facades;
-using InventoryItems.ViewModels;
+using CoinCompanion.Web.Server.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
 
-namespace InventoryItems.Controllers {
+namespace CoinCompanion.Web.Server.Controllers {
     [Route("api/[controller]")]
     [ApiController]
     public class CollectionsController : ControllerBase
