@@ -2,7 +2,8 @@ using Autofac.Extensions.DependencyInjection;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
-namespace InventoryItems {
+namespace CoinCompanion.Web.Server
+{
     public class Program
     {
         public static void Main(string[] args)
