@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace InventoryItems.ViewModels {
+namespace CoinCompanion.Web.Server.ViewModels {
     public class CollectionViewModel {
         public string Name { get; set; }
         public Guid Id { get; set; }
